frc.Robot now dispatches basic method invocations to a collection of RobotLikeComponent objects.
This is grossly incomplete in several ways.  First, there's no way to add
such a component to a Robot.  Second, there's no concrete class implementing
the RobotLikeComponent interface yet.

The frc.Robot class also currently contains a lot of code which needs to
be moved to such a component.

Still, this is a reasonable breakpoint for this effort.