Oops.  Forgot that Iterable lacked an add() method.  Use Collection. Note that I'm trying to use the most abstract type possible.  This is always a good idea, leaving the code maximally flexible.
At the moment, I'm assuming (in the Robot constructor) that the collection
will be a set.  But, if I later learn that (perhaps only some of the time)
order is important, then I can switch this to a List-y type w/o breaking
anything else.  In fact, it can be a runtime choice whether to use a Set-y
type of type or a List-y type of time (ie. whether or not the collection
will preserve order).