Drivetrain now implements RobotLikeComponent.  This required some renaming of methods in Drivetrain to match the standard for the FRC robotic class, but even absent the goal of my current effort, that's a good idea if only for consistency and ease of understanding of the code.
This also required extending the RobotLikeComponent to accept an optional
"debug" argument to methods.  The current default is false.

Should it really use an argument?  It seems to me that there may be a
more efficient approach, such as checking some "registry" or set of
options that can be configured at runtime.  In more conventional code,
I could simply use environment variables.